refactor: reconstruct the date-related modules
