+++ tests/unittest/test_date_utils.py
@@ -7,12 +7,13 @@ from delorean.date_utils import (
     DateSpan,
     SpanGranularity,
 )
+from delorean.date_utils.date_granularity import Daily
 
 
 class DateGranularityTestCase(TestCase):
 
     def test_enum_items(self):
-        self.assertEqual(DateGranularity.DAILY.value, 'daily')
+        self.assertIsInstance(DateGranularity.DAILY.value, Daily)
 
     def test_validate_daily_completion(self):
         start_date = datetime.date(2024, 6, 10)
